backend/requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
psycopg2-binary>=2.9.10
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-jose>=3.3.0
python-dotenv>=1.0.1
python-multipart>=0.0.20
